(feat) fix double images pasted instead of one

+++ ssu_app/public/assets/js/screenshot.js
@@ -1,52 +1,56 @@
 document.addEventListener('DOMContentLoaded', () => {
     const pasteArea = document.getElementById('pasteArea')
+    foundImage = false
 
     pasteArea.addEventListener('paste', (e) => {
         e.stopPropagation()
         e.preventDefault()
-        var clipboardElement = e.clipboardData.items
-        for (item of clipboardElement) {
-            if (item.kind === 'file') {
-                var imageFile = item.getAsFile()
-                var reader = new FileReader()
-                reader.onload = function (event) {
-                    event.preventDefault()
-                    const imgWrapper = document.createElement('div')
-                    imgWrapper.classList.add('image-wrapper')
+        if (!foundImage) {
+            var clipboardElement = e.clipboardData.items
+            for (item of clipboardElement) {
+                if (item.kind === 'file') {
+                    foundImage = true
+                    var imageFile = item.getAsFile()
+                    var reader = new FileReader()
+                    reader.onload = function (event) {
+                        event.preventDefault()
+                        const imgWrapper = document.createElement('div')
+                        imgWrapper.classList.add('image-wrapper')
 
-                    var blob = new Blob([event.target.result])
-                    var imageURL = URL.createObjectURL(blob)
+                        var blob = new Blob([event.target.result])
+                        var imageURL = URL.createObjectURL(blob)
 
-                    const img = document.createElement('img')
-                    img.src = imageURL
-                    const hiddenfileid = createNewIdFileName(imageURL)
-                    console.log("imageURL")
-                    console.log(hiddenfileid)
+                        const img = document.createElement('img')
+                        img.src = imageURL
+                        const hiddenfileid = createNewIdFileName(imageURL)
 
-                    const removeBtn = document.createElement('button')
-                    const filenameOnsrv = document.createElement('input')
-                    filenameOnsrv.setAttribute('type', 'hidden')
-                    filenameOnsrv.setAttribute('id', hiddenfileid)
-                    removeBtn.innerText = 'X'
-                    removeBtn.setAttribute('id', 'btn_' + hiddenfileid)
-                    removeBtn.classList.add('remove-btn')
-                    removeBtn.addEventListener('click', (e) => {
-                        e.preventDefault()
-                        const filenamehidden = document.getElementById(hiddenfileid)
-                        removeFile(filenamehidden.value)
-                        imgWrapper.remove()
-                        console.log("triggered")
-                    })
+                        const removeBtn = document.createElement('button')
+                        const filenameOnsrv = document.createElement('input')
+                        filenameOnsrv.setAttribute('type', 'hidden')
+                        filenameOnsrv.setAttribute('id', hiddenfileid)
+                        removeBtn.innerText = 'X'
+                        removeBtn.setAttribute('id', 'btn_' + hiddenfileid)
+                        removeBtn.classList.add('remove-btn')
+                        removeBtn.addEventListener('click', (e) => {
+                            e.preventDefault()
+                            const filenamehidden = document.getElementById(hiddenfileid)
+                            removeFile(filenamehidden.value)
+                            imgWrapper.remove()
+                        })
 
-                    imgWrapper.appendChild(img)
-                    imgWrapper.appendChild(filenameOnsrv)
-                    imgWrapper.appendChild(removeBtn)
-                    pasteArea.appendChild(imgWrapper)
-                    submitFileForm(event.target.result, 'paste', hiddenfileid)
+                        imgWrapper.appendChild(img)
+                        imgWrapper.appendChild(filenameOnsrv)
+                        imgWrapper.appendChild(removeBtn)
+                        pasteArea.appendChild(imgWrapper)
+                        submitFileForm(event.target.result, 'paste', hiddenfileid)
+                    }
+                    reader.readAsArrayBuffer(imageFile)
                 }
-                reader.readAsArrayBuffer(imageFile)
             }
         }
+        setTimeout(function () {
+            foundImage = false
+        }, 2000)
     })
 })
 
@@ -67,7 +71,7 @@ function submitFileForm(file, type, hiddenfileid) {
     formData.append('elementNumber', elementNumber)
 
     let myBlob = new Blob([file], { "type": "image/png" })
-    formData.append('file', myBlob, elementDir + filename +'-file-' + elementNumber + '.png')
+    formData.append('file', myBlob, elementDir + filename + '-file-' + elementNumber + '.png')
 
     let uploaderPath = document.getElementById('uploaderPath').value
 
@@ -86,7 +90,7 @@ function submitFileForm(file, type, hiddenfileid) {
         } else if (xhr.status == 500) {
             msgElement.innerText = response.data.msg
             msgElement.classList.add(response.data.class)
-        } else  if (xhr.status == 401) {
+        } else if (xhr.status == 401) {
             msgElement.innerText = response.data.msg
             msgElement.classList.add(response.data.class)
         }
@@ -102,8 +106,6 @@ function removeFile(hiddenFilename) {
     formData.append('filename', filename)
     formData.append('elementDir', elementDir)
     formData.append('hiddenFilename', hiddenFilename)
-    console.log("hiddenFilename")
-    console.log(formData)
 
     var xhr = new XMLHttpRequest()
     xhr.open('POST', uploaderPath + '/removeFile', true)
